+++ src/browserist/browser/combo/__main__.py
@@ -13,14 +13,14 @@ class ComboDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver) -> None:
         super().__init__(browser_driver)
 
-    def cookie_banner(self, settings: CookieBannerSettings, timeout: int | None = None) -> None:
+    def cookie_banner(self, settings: CookieBannerSettings, timeout: float | None = None) -> None:
         """Standardised combination of methods to accept or decline cookies."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             combo_cookie_banner(self, settings, timeout)
 
-    def log_in(self, login_credentials: LoginCredentials, login_form: LoginForm1Step | LoginForm2Steps, timeout: int | None = None) -> None:
+    def log_in(self, login_credentials: LoginCredentials, login_form: LoginForm1Step | LoginForm2Steps, timeout: float | None = None) -> None:
         """Standardised combination of methods to log in.
 
         wait_seconds: Extra seconds in addition to timeout to make sure the login is processed and that the user is redirected succesfully."""
@@ -29,7 +29,7 @@ class ComboDriverMethods(DriverMethods):
             timeout = self._mediate_timeout(timeout)
             combo_log_in(self, login_credentials, login_form, timeout)
 
-    def search(self, term: str, settings: SearchSettings, timeout: int | None = None) -> None:
+    def search(self, term: str, settings: SearchSettings, timeout: float | None = None) -> None:
         """Standardised combination of methods to perform search."""
 
         if self._timeout_should_continue():
